Fix bug #50:Calcutator has been rewritten

+++ src/main/java/org/mihalis/opal/calculator/Calculator.java
@@ -140,12 +140,12 @@ public class Calculator extends Composite {
 
 	/**
 	 * @param value new value
-	 * @throws NumberFormatException if <code>value</code> is not a valid float
+	 * @throws NumberFormatException if <code>value</code> is not a valid double
 	 *             value
 	 */
 	public void setValue(final String value) {
 		checkWidget();
-		new Float(value);
+		new Double(value);
 		this.displayArea.setText(value);
 	}
